+++ changelog.md
@@ -222,7 +222,7 @@
 
 - Added `os.getCacheDir()` to return platform specific cache directory.
 
-- Added `os.parseShellCommand` such that `parseShellCommand(quoteShellCommand(a)) == a`.
+- Added `os.splitShellCmd` such that `splitShellCmd(quoteShellCommand(a)) == a`.
 
 - Added a simpler to use `io.readChars` overload.